Implement DetailViewModel
