+++ app/src/main/java/com/skydoves/disneycompose/ui/details/DetailViewModel.kt
@@ -1,0 +1,45 @@
+/*
+ * Designed and developed by 2020 skydoves (Jaewoong Eum)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.skydoves.disneycompose.ui.details
+
+import androidx.annotation.WorkerThread
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.skydoves.disneycompose.base.LiveCoroutinesViewModel
+import com.skydoves.disneycompose.model.Poster
+import com.skydoves.disneycompose.repository.DetailRepository
+import dagger.hilt.android.lifecycle.HiltViewModel
+import timber.log.Timber
+import javax.inject.Inject
+
+@HiltViewModel
+class DetailViewModel @Inject constructor(
+  private val detailRepository: DetailRepository
+) : LiveCoroutinesViewModel() {
+
+  private var _posterDetails: LiveData<Poster> = MutableLiveData()
+  val posterDetails: LiveData<Poster> get() = _posterDetails
+
+  init {
+    Timber.d("init DetailViewModel")
+  }
+
+  @WorkerThread
+  fun getPoster(id: Long) {
+    _posterDetails = detailRepository.getPosterById(id)
+  }
+}
