+++ app/src/main/java/com/skydoves/disneycompose/ui/main/Main.kt
@@ -27,6 +27,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navArgument
 import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.insets.ProvideWindowInsets
+import com.skydoves.disneycompose.ui.details.DetailViewModel
 import com.skydoves.disneycompose.ui.details.PosterDetails
 import com.skydoves.disneycompose.ui.posters.Posters
 
@@ -55,7 +56,7 @@ fun DisneyMain() {
           navArgument(NavScreen.PosterDetails.argument0) { type = NavType.LongType }
         )
       ) { backStackEntry ->
-        val viewModel = hiltViewModel<MainViewModel>(backStackEntry = backStackEntry)
+        val viewModel = hiltViewModel<DetailViewModel>(backStackEntry = backStackEntry)
 
         val posterId =
           backStackEntry.arguments?.getLong(NavScreen.PosterDetails.argument0) ?: return@composable
