+++ app/src/main/java/com/skydoves/disneycompose/ui/main/MainViewModel.kt
@@ -18,7 +18,6 @@ package com.skydoves.disneycompose.ui.main
 
 import androidx.annotation.MainThread
 import androidx.annotation.StringRes
-import androidx.annotation.WorkerThread
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
@@ -29,7 +28,6 @@ import androidx.lifecycle.switchMap
 import coil.ImageLoader
 import com.skydoves.disneycompose.base.LiveCoroutinesViewModel
 import com.skydoves.disneycompose.model.Poster
-import com.skydoves.disneycompose.repository.DetailRepository
 import com.skydoves.disneycompose.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import timber.log.Timber
@@ -38,15 +36,11 @@ import javax.inject.Inject
 @HiltViewModel
 class MainViewModel @Inject constructor(
   private val mainRepository: MainRepository,
-  private val detailRepository: DetailRepository,
   val imageLoader: ImageLoader
 ) : LiveCoroutinesViewModel() {
 
   private var _posterList: MutableLiveData<Boolean> = MutableLiveData(true)
   val posterList: LiveData<List<Poster>>
-
-  private var _posterDetails: LiveData<Poster> = MutableLiveData()
-  val posterDetails: LiveData<Poster> get() = _posterDetails
 
   private val _isLoading: MutableLiveData<Boolean> = MutableLiveData(false)
   val isLoading: LiveData<Boolean> get() = _isLoading
@@ -69,11 +63,6 @@ class MainViewModel @Inject constructor(
         ).asLiveData()
       }
     }
-  }
-
-  @WorkerThread
-  fun getPoster(id: Long) {
-    _posterDetails = detailRepository.getPosterById(id)
   }
 
   @MainThread
